requirements.txt
accelerate==0.26.1
addict==2.4.0
aiofiles==23.2.1
aiohttp==3.9.1
aiosignal==1.3.1
altair==5.2.0
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==4.2.0
appdirs==1.4.4
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
autocommand==2.2.2
beautifulsoup4==4.12.3
bleach==6.1.0
blessed==1.20.0
blinker==1.7.0
blis==0.7.11
boto3==1.34.54
botocore==1.34.54
braceexpand==0.1.7
cachetools==5.3.2
catalogue==2.0.10
certifi==2023.11.17
cffi==1.16.0
cfgv==3.4.0
chardet==5.2.0
charset-normalizer 
cheroot==10.0.0
CherryPy==18.9.0
click==8.1.7
clip @ git+https://github.com/openai/CLIP.git@a1d071733d7111c9c014f024669f959182114e33
cloudpathlib==0.16.0
colorama==0.4.6
comm==0.2.1
confection==0.1.4
ConfigArgParse==1.7
contexttimer==0.3.3
contourpy==1.2.0
cryptography==42.0.2
cycler==0.12.1
cymem==2.0.8
dash==2.15.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
datasets==2.14.7
decorator==4.4.2
decord==0.6.0
deepspeed==0.12.6
diffusers==0.16.0
dill==0.3.7
distlib==0.3.8
distro==1.9.0
docker-pycreds==0.4.0
docstring-parser==0.15
docx==0.2.4
easydict==1.9
einops==0.7.0
einops-exts==0.0.4
en-core-web-md @ https://github.com/explosion/spacy-models/releases/download/en_core_web_md-3.7.1/en_core_web_md-3.7.1-py3-none-any.whl#sha256=6a0f857a2b4d219c6fa17d455f82430b365bf53171a2d919b9376e5dc9be032e
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
et-xmlfile==1.1.0
exceptiongroup==1.2.0
executing==2.0.1
fairscale==0.4.4
fastapi==0.109.0
fastjsonschema==2.19.1
feedparser==6.0.11
ffmpy==0.3.1
filelock
flash-attn==2.5.8
Flask==3.0.1
fonttools==4.47.2
frozenlist==1.4.1
fsspec==2023.10.0
ftfy==6.1.3
gdown==5.1.0
gitdb==4.0.11
GitPython==3.1.41
gmpy2
google-ai-generativelanguage==0.4.0
google-api-core==2.16.1
google-auth==2.27.0
google-generativeai==0.3.2
googleapis-common-protos==1.62.0
gpustat==1.1.1
gradio==4.8.0
gradio_client==0.7.1
grpcio==1.60.0
grpcio-status==1.60.0
h11==0.14.0
h5py==3.10.0
hjson==3.1.0
httpcore==1.0.2
httpx==0.26.0
huggingface-hub==0.23.0
icecream==2.1.3
identify==2.5.33
idna
ijson==3.2.3
imageio==2.33.1
imageio-ffmpeg==0.4.9
importlib-metadata==7.0.1
importlib-resources==6.1.1
inflect==7.0.0
inflection==0.5.1
iniconfig==2.0.0
inquirerpy==0.3.4
iopath==0.1.10
ipython==8.18.1
ipywidgets==8.1.1
itsdangerous==2.1.2
jaraco.collections==5.0.0
jaraco.context==4.3.0
jaraco.functools==4.0.0
jaraco.text==3.12.0
jedi==0.19.1
Jinja2
jmespath==1.0.1
joblib==1.3.2
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
jupyter_core==5.7.1
jupyterlab-widgets==3.0.9
kaggle==1.6.4
kiwisolver==1.4.5
langcodes==3.3.0
lazy_loader==0.3
Levenshtein==0.23.0
lmdb==1.4.1
lxml==5.1.0
Markdown==3.5.2
markdown-it-py==3.0.0
markdown2==2.4.12
MarkupSafe
matplotlib==3.8.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mkl-fft
mkl-random
mkl-service==2.4.0
more-itertools==10.2.0
moviepy==1.0.3
mpmath
multidict==6.0.4
multiprocess==0.70.15
murmurhash==1.0.10
natsort==8.4.0
nbformat==5.9.2
nest-asyncio==1.6.0
networkx
ninja==1.11.1.1
nltk==3.8.1
nodeenv==1.8.0
numpy
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py==12.535.133
nvidia-nccl-cu12
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
open-clip-torch==2.24.0
open3d==0.18.0
openai==1.7.2
opencv-python-headless==4.5.5.64
opendatasets==0.1.22
openpyxl==3.1.2
orjson==3.9.10
packaging==23.2
pandas==2.1.4
parso==0.8.3
pattern3==3.0.0
pdfminer.six==20231228
pdfminer3k==1.3.4
peft==0.7.1
pexpect==4.9.0
pfzy==0.3.4
pillow
platformdirs==4.2.0
plotly==5.18.0
pluggy==1.3.0
ply==3.11
portalocker==2.8.2
portend==3.2.0
pre-commit==3.6.0
preshed==3.0.9
prettytable==3.9.0
proglog==0.1.10
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.25.2
psutil==5.9.7
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyarrow==14.0.2
pyarrow-hotfix==0.6
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycocoevalcap==1.2
pycocotools==2.0.7
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
pydeck==0.8.1b0
pydub==0.25.1
Pygments==2.17.2
pynvml==11.5.0
pyparsing==3.1.1
pyquaternion==0.9.9
PySocks==1.7.1
pytest==7.4.4
python-dateutil==2.8.2
python-Levenshtein==0.23.0
python-magic==0.4.27
python-multipart==0.0.6
python-slugify==8.0.3
pytz==2023.3.post1
PyYAML
pyyaml_env_tag==0.1
Random-Word==1.0.11
rapidfuzz==3.6.1
referencing==0.32.1
regex==2023.12.25
requests
retrying==1.3.4
rich==13.7.0
rpds-py==0.17.1
rsa==4.9
s3transfer==0.10.0
safetensors==0.4.1
# Editable install with no version control (salesforce-lavis==1.0.1)
# -e /home/lmm/LAVIS
scikit-image==0.22.0
scikit-learn==1.3.2
scipy==1.11.4
seaborn==0.13.2
semantic-version==2.10.0
sentencepiece==0.1.99
sentry-sdk==1.39.2
setproctitle==1.3.3
sgmllib3k==1.0.0
shellingham==1.5.4
shtab==1.6.5
simplejson==3.19.2
six==1.16.0
smart-open==6.4.0
smmap==5.0.1
sniffio==1.3.0
soundfile==0.12.1
soupsieve==2.5
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
stack-data==0.6.3
starlette==0.35.1
streamlit==1.30.0
sympy
tempora==5.5.0
tenacity==8.2.3
text-unidecode==1.3
thinc==8.2.2
threadpoolctl==3.2.0
tifffile==2024.1.30
tiktoken==0.5.2
timm==0.9.12
tokenizers==0.19.1
toml==0.10.2
tomli==2.0.1
tomlkit==0.12.0
toolz==0.12.0
tornado==6.4
tqdm==4.66.1
traitlets==5.14.1
transformers-stream-generator==0.0.4
triton
trl==0.7.9
typer==0.9.0
typing_extensions 
tyro==0.6.4
tzdata==2023.4
tzlocal==5.2
urllib3==1.26.18
uvicorn==0.25.0
validators==0.22.0
virtualenv==20.25.0
wandb==0.16.2
wasabi==1.1.2
watchdog==3.0.0
wcwidth==0.2.13
weasel==0.3.4
webdataset==0.2.86
webencodings==0.5.1
websockets==11.0.3
Werkzeug==3.0.1
widgetsnbextension==4.0.9
xformers==0.0.24+cu118
XlsxWriter==3.1.9
xxhash==3.4.1
yajl==0.3.5
yarl==1.9.4
zc.lockfile==3.0.post1
zipp==3.17.0
